student_deineka_vladislav/src/main/java/lv/javaguru/travel/insurance/core/validation/LastNameValidator.java
package lv.javaguru.travel.insurance.core.validation;

import lv.javaguru.travel.insurance.dto.TravelCalculatePremiumRequest;
import lv.javaguru.travel.insurance.dto.ValidationErrors;
import org.springframework.stereotype.Component;

import java.util.Optional;

@Component
class LastNameValidator implements TravelRequestValidation{

    @Override
    public Optional<ValidationErrors> execute(TravelCalculatePremiumRequest request) {
        return (request.getPersonLastName() == null || request.getPersonLastName().isEmpty())
                ? Optional.of(new ValidationErrors("lastName", "Field cannot be empty"))
                : Optional.empty();

    }
}